% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prebuilt_wss.R
\name{list_prebuilt_wss}
\alias{list_prebuilt_wss}
\title{List Available Prebuilt WSS Reference Sets}
\usage{
list_prebuilt_wss()
}
\value{
A character vector of prebuilt reference set names usable as the
\code{prebuilt} argument of \code{\link[=calculate_WSS]{calculate_WSS()}} and \code{\link[=load_prebuilt_wss]{load_prebuilt_wss()}}.
}
\description{
List Available Prebuilt WSS Reference Sets
}
\examples{
list_prebuilt_wss()

}
