% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prebuilt_wss.R
\name{wss_prebuilt_terms}
\alias{wss_prebuilt_terms}
\title{Get Module Term Labels for a Prebuilt WSS Reference Set}
\usage{
wss_prebuilt_terms(prebuilt = "core_AD_plasma_biomarkers")
}
\arguments{
\item{prebuilt}{Name of a prebuilt reference set. See
\code{\link[=list_prebuilt_wss]{list_prebuilt_wss()}} for available options.}
}
\value{
A data.table of module term labels, or \code{NULL} if the given
reference set has no associated terms file.
}
\description{
Some prebuilt reference sets ship a companion table of human-readable
term labels for each module, summarized from the underlying pathway
enrichment used to build the module.
}
\examples{
head(wss_prebuilt_terms())

}
